noa-frontend/noa-backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==1.10.13
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pandas==2.2.1
openai==1.14.3
